+++ secrets/drivertest/drivertest.go
@@ -14,15 +14,15 @@
 
 // Package drivertest provides a conformance test for implementations of
 // the secrets driver.
-package drivertest // import "gocloud.dev/internal/secrets/drivertest"
+package drivertest // import "gocloud.dev/secrets/drivertest"
 
 import (
 	"context"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
-	"gocloud.dev/internal/secrets"
-	"gocloud.dev/internal/secrets/driver"
+	"gocloud.dev/secrets"
+	"gocloud.dev/secrets/driver"
 )
 
 // Harness descibes the functionality test harnesses must provide to run
